Int ID移行: ビルドエラー修正完了
- GuildView: jobIndex→jobId、raceOrder辞書型修正
- StoryView: navigationDestination型をUInt16に変更
- ExplorationSnapshot: dungeonIdをUInt16に統一
- GemModificationProgressService: xxxIndex→xxxId命名統一
- InventoryProgressService: 型変換追加（UInt8/UInt16/Int）
- ItemPreloadService: xxxIndex→xxxId命名統一
- DropModels: SuperRareDailyState.jstDateをUInt32に変更
- ShopProgressService: 未使用変数警告修正
- 各種サービス・ビュー層の型整合性修正

ビルド成功（警告0件）

+++ Epika/Progress/Application/Services/ExplorationProgressService.swift
@@ -16,7 +16,7 @@ final class ExplorationProgressService {
     }
 
     private enum ExplorationSnapshotBuildError: Error {
-        case dungeonNotFound(UInt8)
+        case dungeonNotFound(UInt16)
         case itemNotFound(UInt16)
         case enemyNotFound(UInt16)
         case statusEffectNotFound(String)
